nineteen update

+++ src/app/components/teams/teams.component.ts
@@ -7,6 +7,7 @@ import { EquipeService } from 'src/app/services/equipe.service';
 import { Equipe } from 'src/app/models/equipe';
 import { TokenStorageService } from 'src/app/services/token-storage.service';
 import { Router } from '@angular/router';
+import { FormBuilder, FormGroup } from '@angular/forms';
 
 
 @Component({
@@ -15,8 +16,10 @@ import { Router } from '@angular/router';
   styleUrls: ['./teams.component.css']
 })
 export class TeamsComponent implements OnInit {
+  formValue : FormGroup;
 
   constructor(
+    private formBuilder : FormBuilder,
     private equipeService : EquipeService,
     private token : TokenStorageService ,
     private router:Router ) { }
@@ -35,6 +38,10 @@ export class TeamsComponent implements OnInit {
   }
 
   ngOnInit(): void {
+    this.formValue = this.formBuilder.group({
+      nom : [''],
+      objectif :['']
+    })
     this.currentUser = this.token.getUtilisateur();
     this.userPermission = this.permissions();
     if (this.userPermission) {
@@ -42,7 +49,10 @@ export class TeamsComponent implements OnInit {
       this.dataSource.sort = this.sort;
       this.dataSource.paginator = this.paginator;
   }
-
+onEdit(row : any){
+  this.formValue.controls['nom'].setValue(row.nom);
+  this.formValue.controls['objectif'].setValue(row.objectif);
+}
 goTo(){
   this.router.navigate([
     '/addteams'
@@ -50,10 +60,7 @@ goTo(){
   ])
 
 }
-goEdit(id: any) {
-  this.router.navigate([`'editTeam/${id}`]);
 
-}
   getEquipes() : void
   {
     this.equipeService.getEquipes().subscribe(
@@ -68,6 +75,8 @@ goEdit(id: any) {
   }
 
   public deleteEquipe(id : number): void{
+    let conf = confirm("Are you sure?");
+    if(!conf) return;
     this.equipeService.deleteEquipe(id).subscribe(
       (response: void) => {
         console.log(response);
